requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# LangChain & OpenAI
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
langchain-text-splitters==0.2.0

# Vector Database
qdrant-client==1.9.0

# Slack Integration
slack-bolt==1.18.1

# Document Loading
pypdf==4.2.0
python-docx==1.1.2
unstructured==0.14.4
tiktoken==0.7.0

# Configuration
python-dotenv==1.0.1